new try

+++ hashmap.c
@@ -82,7 +82,7 @@ void eraseMap(HashMap * map,  char * key) {
 
 Pair * searchMap(HashMap * map,  char * key)
 {   
-  int index=hash(key,map->capacity)
+  long index=hash(key,map->capacity)
   while(map->buckets[index]!=NULL)
     {
       if(is_equal(key,map->buckets[index]->key)==1)
